% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{calclmodp}
\alias{calclmodp}
\title{Calculate linear model and p-value}
\usage{
calclmodp(x, df)
}
\arguments{
\item{x}{string of gene name}

\item{df}{data frame of subject variance rate of change}
}
\value{
linear model coefficients
}
\description{
This function takes a gene symbol and a subject variance rate of change data frame
as input to subseting it for the specified gene. It then fits a linear regression
estimating the intercept and slope of subject variance and a p-value of significance
for the fit. These coefficients are returned as an output
}
