% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{compute_multisignature_stability}
\alias{compute_multisignature_stability}
\title{Compute Multi-Signature Stability}
\usage{
compute_multisignature_stability(signatures)
}
\arguments{
\item{signatures}{named list of vectors of gene symbols}
}
\value{
A data frame of subject variance per signature and celltype
}
\description{
This function computes the average single cell subject variance for a gene signature.
Input is a vector of gene symbols and it retrieves the distribution of its subject
variance to the distribution of all genes within each celltype as background.
}
