% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{compute_signature_stability}
\alias{compute_signature_stability}
\title{Compute Signature Stability}
\usage{
compute_signature_stability(signature)
}
\arguments{
\item{signature}{vector of gene symbols}
}
\value{
A data frame of subject variance per gene and celltype
}
\description{
This function computes the average single cell subject variance for a gene signature.
Input is a vector of gene symbols and it retrieves the distribution of its subject
variance to the distribution of all genes within each celltype as background.
}
