% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{plot_signature_gene_agetrend}
\alias{plot_signature_gene_agetrend}
\title{Plot rateOfchange of signature subject variance per gene}
\usage{
plot_signature_gene_agetrend(signature_rateOfchange, cell_subsets)
}
\arguments{
\item{signature_rateOfchange}{data frame of subject variance rate of change for gene signature}

\item{cell_subsets}{celltypes to be visualized}
}
\value{
lineplot
}
\description{
This function takes the data frame of subject variance rate of change for a gene
signature as input and visualizes the subject variance trend for each gene per
celltype as a linear trend lines.
}
