% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{plot_signature_stability_bars}
\alias{plot_signature_stability_bars}
\title{Plot Stacked Bars of Signature Variance Rate of Change Compostion per Celltype}
\usage{
plot_signature_stability_bars(signature_stability_slope)
}
\arguments{
\item{signature_stability_slope}{data frame of subject variance slope of rate of change}
}
\value{
barplot
}
\description{
This function takes a data frame of the slope of subject variance per gene of
a gene signature and celltype as input and categorizes each gene into increasing
(positive slope, p<=0.05) or decreasing (negative slope, p<=0.05) or stable
(p>0.05). It then visualizes the fractions of these categories as radarplot per
celltype.
}
