% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{plot_signature_stability_heatmap}
\alias{plot_signature_stability_heatmap}
\title{Heatmap of Subject Variance per Signature}
\usage{
plot_signature_stability_heatmap(multisignature_stability_df)
}
\arguments{
\item{A}{data frame of subject variance per signature and celltype}
}
\value{
heatmap
}
\description{
This function takes a vector of signature names as input and
visualizes the distribution of its subject variance in contrast
to the background in a violinplot.
}
