% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{compute_signature_stability_rateofchange}
\alias{compute_signature_stability_rateofchange}
\title{Calculate Rate of Change Subject Variance}
\usage{
compute_signature_stability_rateofchange(signature_name, signature)
}
\arguments{
\item{signature_name}{string of signature name}

\item{signature}{vector of gene symbols}
}
\value{
data frame of subject variance rate of change per gene and celltype
}
\description{
This function takes a signature name and a vector of signature gene symbols as
input and retrieves the estimated subject variance in sliding age windows of
size 3 in bulk and size 5 in single cell leading to the rate of change across
age. It returns a data frame of mean and median subject variance per window,
gene and celltype.
}
