% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{compute_signature_stability_slope}
\alias{compute_signature_stability_slope}
\title{Compute subject variance slope for a signature per celltype}
\usage{
compute_signature_stability_slope(signature_name, signature_gene_list)
}
\arguments{
\item{signature_name}{string of signature name}

\item{signature_gene_list}{vector of gene symbols}
}
\value{
data frame of subject variance slope of rate of change
}
\description{
This function takes a signature name and a vector of signature gene symbols as
input and retrieves the estimated slope of subject variance across age windows.
It returns a data frame of the slope of subject variance per window gene and
celltype.
}
