% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{plot_signature_agetrend}
\alias{plot_signature_agetrend}
\title{Plot the Average Rate of Change of Signature Subject Variance per Celltype}
\usage{
plot_signature_agetrend(signature_rateOfchange, cell_subsets)
}
\arguments{
\item{signature_rateOfchange}{data frame of subject variance rate of change for gene signature}

\item{cell_subsets}{celltypes to be visualized}
}
\value{
lineplot
}
\description{
This function takes the data frame of subject variance rate of change for a gene
signature as input and visualizes the average subject variance trend across all
genes per celltype as a loess spline fit.
}
