% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{plot_signature_stability}
\alias{plot_signature_stability}
\title{Plot Violins of Subject Variance}
\usage{
plot_signature_stability(signature_name, signature_stability_df)
}
\arguments{
\item{signature_name}{string of signature name}

\item{signature_stability_df}{data frame of subject variance per gene and celltype}
}
\value{
violinplot
}
\description{
This function takes a signature name and single cell subject variance data frame as
input and visualizes the distribution of its subject variance in contrast to the
background in a violinplot.
}
